Move snackbar to the `shared/ui` folder
