--- src/lib/features/snackbar/index.ts
@@ -1,3 +1,0 @@
-export type { SnackbarOptions } from './model/types';
-export { snackbar, snackbarQueue } from './model/store';
-export { default as SnackbarQueue } from './ui/snackbar-queue.svelte';
