+++ src/lib/shared/ui/snackbar/index.ts
@@ -1,0 +1,3 @@
+export type { SnackbarOptions } from './types';
+export { snackbar, snackbarQueue } from './store';
+export { default as SnackbarQueue } from './snackbar-queue.svelte';
