+++ src/lib/shared/ui/snackbar/snackbar-queue.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { snackbarQueue } from '../model/store';
+  import { snackbarQueue } from './store';
   import Snackbar from './snackbar.svelte';
 
   $: activeSnackbar = $snackbarQueue[0] || null;
