+++ src/lib/shared/ui/snackbar/snackbar.spec.ts
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/svelte';
 import userEvent from '@testing-library/user-event';
-import { SnackbarQueue, snackbar, type SnackbarOptions } from '..';
+import { SnackbarQueue, snackbar, type SnackbarOptions } from '.';
 
 const firstSnackbar: SnackbarOptions = {
   text: 'Hello',
